Port qmstr to use Cobra, shift basic functionality to qmstr run.
References #286.
